Add the Map trait.
The Map trait makes use of a type alias around a Boxed emitter. We use a
Box and not a reference because we want the Mapper to own the emitter
such that it goes out of scope once the map function has finished. This
saves the user of Map::map from managing the emitter's lifetime
themselves.

Added test showing associated types.

Also removed the option for the implementation of Map to change the
input type. It is now always Into<String>.