Change map and reduce to instance methods.
The payload runner functions need to know where the user's Map and
Reduce implementations are. The only way to do this is to pass
references to instances of these structs to a registration function. As
such, map and reduce need to be changed to instance methods instead of
static functions.