Add task restarting to master.
The master now tracks how many times a worker has failed to respond to a
poll. After 5 times, the master will unschedule the task assigned to the
worker and place it back on the queue. After 10 times, the master will
remove the worker from the list of active workers and will no longer
poll it.