Remove master integration test.
The integration test fails if the master tests are run from a clean
build environment. The shell script has been left in since it can be
repurposed, but it no longer runs automatically as part of `cargo test`.
Also fix errors in the shell script and set it to fail on error.