Add basic cluster dashboard
This commit adds the basic structure and functionality for the master
dashboard. Currently this dashboard just shows the current information
for the cluster and can not be interacted with. The design of the
dashboard is also very basic but can be improved upon in future commits.

The dashboard is served by the master using the Iron web framework.